[modify] vote: component、邏輯拆分、RWD 及切版ver1 完成
